move integration test into appropriate packages
