+++ feature/integrationtest/src/main/java/com/composum/platform/features/integrationtest/SmokeIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.sling.launchpad;
+package com.composum.platform.features.integrationtest;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.notNullValue;
