+++ feature/integrationtest/src/main/java/com/composum/platform/features/integrationtest/UrlCheck.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.sling.launchpad;
+package com.composum.platform.features.integrationtest;
 
 import java.net.ConnectException;
 import java.util.Collection;
